committing header in all pages and some ui changes

+++ src/main/webapp/js/jquery/domainManager.js
@@ -2,6 +2,7 @@ function getURLParameter(name) {
 	return decodeURI((RegExp(name + '=' + '(.+?)(&|$)').exec(location.search) || [
 			, null ])[1]);
 }
+
 var requestedUserId = "null";
 
 $(function() {
@@ -12,10 +13,9 @@ $(function() {
 				url : 'domains?key=0',
 				success : function(data) {
 					data = JSON.parse(data);
-					// alert(data.length);
 					for (i = 0; i < data.length; i++) {
 						var table = "";
-						table += "<table style='margin-top:15px;margin-left:20px;width:100%;' class='imagetable' id='tableId"
+						table += "<table style='margin-top:15px;width:100%;' class='imagetable' id='tableId"
 								+ i + "'>";
 						table += "<tr pid='0' id="
 								+ data[i].key
@@ -23,8 +23,8 @@ $(function() {
 										"<td class='iconWidth'><img class='newDomain' src='images/new.png'/></td>" +
 										"<td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td>" +
 										"<td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td>" +
-										"<td class='titleClass' class='addBorder' ><input   readonly='readonly' " +
-										"class='addBorder autoCompleteWiki' type='text' value='"+ data[i].title +"'>" +
+										"<td class='titleClass' class='addBorder' ><span class='spanTitle'>"+data[i].title+"</span><input style='display:none' " +
+										"class='autoCompleteWiki' type='text' value='"+ data[i].title +"'>" +
 												"</td><td></td><td></td><td></td></tr>"
 						$('#domainDivId').append(table);
 					}
@@ -34,6 +34,7 @@ $(function() {
 				dataType : 'text'
 			});
 
+	
 });
 
 function showCompleteDomainTree(rootKey, table) {
@@ -63,9 +64,10 @@ function showCompleteDomainTree(rootKey, table) {
 
 								// alert(ratingStarHtml);
 								ul += "<tr pid='"+rootKey+"' id='"+obj.key+"' "
-										+ "><td colspan='5' style='border-top:none;border-bottom:none;'></td><td class='iconWidth'><img class='newDomain' src='images/new.png'/></td><td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td style='border-bottom:none'><input   readonly='readonly' class='addBorder autoCompleteWiki' type='text' value='"
+										+ "><td colspan='5' style='border-top:none;border-bottom:none;'></td><td class='iconWidth'><img class='newDomain' src='images/new.png'/></td><td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td class='titleClass'><span class='spanTitle'>"+obj.title+"</span> " +
+												"<input style='display:none'  class='autoCompleteWiki' type='text' value='"
 										+ obj.title
-										+ "'></td><td><input readonly='readonly' class='inputWeightage addBorder' type='text' value="
+										+ "'></td><td><input  readonly='readonly' class='inputWeightage addBorder' type='text' value="
 										+ obj.weightage
 										+ "></td><td></td><td score="
 										+ obj.score + " id=p" + i + ">"
@@ -79,11 +81,11 @@ function showCompleteDomainTree(rootKey, table) {
 									weightage = obj1.weightage;
 									key = obj1.key;
 
-									ul += "<tr pid='"+obj.key+"' id='"+obj1.key+"'><td colspan='11' style='border-top:none;border-bottom:none;'></td><td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td  i="
+									ul += "<tr pid='"+obj.key+"' id='"+obj1.key+"'><td colspan='11' style='border-top:none;border-bottom:none;'></td><td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td class='titleClass'  i="
 											+ i
 											+ " j="
 											+ j
-											+ "><input  readonly='readonly' class='addBorder autoCompleteWiki' type='text' value='"
+											+ "><span class='spanTitle'>"+obj1.title+"</span><input style='display:none' class='autoCompleteWiki' type='text' value='"
 											+ obj1.title
 											+ "'></td><td><input  readonly='readonly' class='addBorder inputWeightage' type='text' value="
 											+ obj1.weightage + "></td></tr>";
@@ -92,9 +94,9 @@ function showCompleteDomainTree(rootKey, table) {
 								// ul += "<tr></tr>";
 							} else {
 
-								ul += "<tr pid='"+rootKey+"' id='"+obj.key+"' childCount=0><td colspan='5' style='border-top:none;border-bottom:none;'></td><td class='iconWidth'><img class='newDomain' src='images/new.png'/></td><td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td><input  readonly='readonly' class='addBorder autoCompleteWiki' type='text' value='"
+								ul += "<tr pid='"+rootKey+"' id='"+obj.key+"' childCount=0><td colspan='5' style='border-top:none;border-bottom:none;'></td><td class='iconWidth'><img class='newDomain' src='images/new.png'/></td><td class='iconWidth'><img class='editDomain' src='images/edit.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td class='titleClass'><span class='spanTitle'>"+obj.title+"</span><input style='display:none' class='autoCompleteWiki' type='text' value='"
 										+ obj.title
-										+ "'></td><td><input  readonly='readonly' class='inputWeightage addBorder'  type='text' value="
+										+ "'></td><td><input   class='inputWeightage addBorder'  type='text' value="
 										+ obj.weightage
 										+ "></td><td></td><td assessmentId="
 										+ obj.assessmentId
@@ -119,65 +121,69 @@ function showCompleteDomainTree(rootKey, table) {
 
 
 $(function() {
-	$('td').live(
-			'click',
-			function() {
-				var colIndex = $(this).parent().children().index($(this));
-				var rowIndex = $(this).parent().parent().children().index(
-						$(this).parent());
-				// alert('Row: ' + rowIndex + ', Column: ' + colIndex);
-			});
+	
 
 	$(".editDomain").live('click', function() {
 		var row = $(this).closest('tr');
-		row.find('input').removeClass("addBorder");
-		row.find('input').attr('readonly',false);
-//		row.find('.titleClass').attr('contentEditable','true');
-//		row.find('.titleClass').css('box-shadow','1px 1px 1px 0 lightgray inset');
-//		row.find('.titleClass').css('border','1px solid gray');
-		// $(this).closest('td img .deleteDomain').hide();
+		var titleInput = row.find('.autoCompleteWiki');
+		titleInput.show();
+		var spanTitle = row.find('.spanTitle');
+		spanTitle.hide();
+		row.find('.inputWeightage').removeClass("addBorder");
 		$(this).attr("src", "images/save.png");
 		$(this).parent().prev().children().hide();
 		$(this).removeClass("editDomain");
 		$(this).addClass("saveDomain");
-//		$(this).parent().next().children().addClass("saveDomain");
-//		$(this).parent().next().children().removeClass("deleteDomain");
+
 	});
 
 	$(".saveDomain").live('click', function() {
-		
-		
-		
 		var row = $(this).closest('tr');
-		row.find('input').attr('readonly',true);
-//		row.find('.titleClass').attr('contentEditable','false');
-//		row.find('.titleClass').css('box-shadow','');
-//		row.find('.titleClass').css('borderow.find('.titleClass').attr('contentEditable','false');
-//		row.find('.titleClass').css('box-shadow','');
-//		row.find('.titleClass').css('border','');r','');
+		var titleInput = row.find('.autoCompleteWiki');
+		var title = titleInput.val();
+		var spanTitle = row.find('.spanTitle');
+		spanTitle.html(title);
+		spanTitle.show();
+		titleInput.hide();
+		var weightageTd = row.find('.inputWeightage');
+		var weightage = weightageTd.val();
+
 		var key = row.attr("id");
+		if(key =='new'){
+			key='0';
+		}
 		var parentKey = row.attr("pid");
 		
 		
 		var table = $(this).closest('table');
-//		var rows = table.find("tr[pid='"+parentKey+"']");
-//		sum=0;
-//		rows.each(function(index) {
-//		    alert($(this).find('input::nth-child(2)').val());
-//		    alert(sum);
-//		});
+		var rows = table.find("tr[pid='"+parentKey+"']");
+		sum=0;
 		
-		var titleTd = row.find('input::nth-child(1)');
-		var title = titleTd.val();
-		var weightageTd = row.find('input::nth-child(2)');
-		var weightage = weightageTd.val();
-		updateDomain(key,parentKey,title,weightage,row);
-		row.find('input').addClass("addBorder");
-		row.find('input').attr('readonly', 'readonly');
+		rows.each(function(index) {
+		    sum = sum + parseInt($(this).find('.inputWeightage').val());
+		});
+		
+		if(sum > 100){
+			alert("Total sum of subdomain's weightage under a domain can't be greater than 100");
+			return;
+		}
+		
+		//row.find('input').attr('readonly',true);
+		
+		
+		saveDomain(key,parentKey,title,weightage,row);
+		row.find('.inputWeightage').addClass("addBorder");
 		$(this).attr("src", "images/edit.png");
 		$(this).addClass("editDomain");
 		$(this).parent().prev().children().show();
 		$(this).removeClass("saveDomain");
+		
+		var expandImg =row.find('.expandDomain');
+		
+		if (expandImg.is(":hidden")) {
+			expandImg.show();  // hide button
+		 }
+		
 	});
 
 	$(".expandDomain").live('click', function() {
@@ -204,8 +210,8 @@ $(function() {
 			var ppid =row.attr("pid");
 			var pid= row.attr("id");
 			if(ppid=='0'){
-				expandDomain =row.find('.expandDomain');
-				if(expandDomain.length>0){
+				expandDomain =row.find('.expandDomain');   
+				if(expandDomain.length>0){						//if domain not expanded then first expand it
 					var table = $(this).closest('table');
 					showCompleteDomainTree(pid, table);
 					expandDomain.attr("src", "images/collapse.png");
@@ -216,50 +222,89 @@ $(function() {
 			
 			var colspan = cell.prev().attr("colspan");
 			if (colspan == 5) {
-				var tr = "<tr id='0' pid='"+pid+"'><td colspan='11'><td><img class='saveDomain' src='images/save.png'></td><td  class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td><input class='autoCompleteWiki' type='text'></td><td><input class='inputWeightage' type='text'></td></tr>";
-				row.after(tr);
-				} else {
-				var tr = "<tr id='0' pid='"+pid+"'><td colspan='5'><td  class='iconWidth'><img style='display:none'  class='newDomain' src='images/new.png'/></td><td><img class='saveDomain' src='images/save.png'></td><td  class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td><input class='autoCompleteWiki' type='text'></td><td><input class='inputWeightage' type='text'></td></tr>";
-				row.after(tr);
-				}
+							var tr = "<tr id='new' pid='"+pid+"'><td class='iconWidth' colspan='11'></td>" +
+							"<td class='iconWidth'><img class='saveDomain' src='images/save.png'></td>" +
+							"<td  class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td>" +
+							"<td class='titleClass'><span class='spanTitle'></span><input class='autoCompleteWiki' type='text'></td>" +
+							"<td><input class='inputWeightage' type='text'></td></tr>";
+							row.after(tr);
+						} else {
+							var tr = "<tr id='new' pid='"
+									+ pid
+									+ "'><td colspan='5'></td><td  class='iconWidth'><img style='display:none'  class='newDomain' src='images/new.png'/></td>" +
+											"<td class='iconWidth'><img class='saveDomain' src='images/save.png'></td>" +
+											"<td  class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td>" +
+											"<td class='titleClass'><span class='spanTitle'></span><input class='autoCompleteWiki' type='text'></td>" +
+											"<td><input class='inputWeightage' type='text'></td></tr>";
+							row.after(tr);
+						}
 			});
 	
 	$(".deleteDomain").live('click', function() {
+		
+		if(!confirm('Are you sure you want to delete domain?'))
+		{
+			return;
+		}
+		
 		var row = $(this).closest('tr');
 		var key = row.attr("id");
 		var parentKey = row.attr("pid");
-		deleteDomain(key,parentKey);
-		//row.remove();
-		var table = $(this).closest('table');
-		if(parentKey != '0'){
-		//var tr = table.find('tr::nth-child(1)');
-		//var key = tr.attr('id');
-		
-		row.remove();
-		table.find("tr[pid='"+key+"']").remove();
-		//showCompleteDomainTree(key, table);
-		$(this).attr("src", "images/collapse.png");
-		$(this).removeClass("expandDomain");
-		$(this).addClass("collapseDomain");
+		if (key != 'new') {
+			deleteDomain(key, parentKey);
 		}
-		else{
-		table.remove();	
+		
+		var table = $(this).closest('table');
+		if (parentKey != '0') {
+			row.remove();
+			table.find("tr[pid='" + key + "']").remove();
+			// showCompleteDomainTree(key, table);
+			$(this).attr("src", "images/collapse.png");
+			$(this).removeClass("expandDomain");
+			$(this).addClass("collapseDomain");
+		} else {
+			table.remove();
 		}
 	});
+	
+
+		$(".inputWeightage").live(
+			'keydown',
+			function(evt) {
+				{
+					var charCode = evt.keyCode;
+
+					if ((charCode > 47 && charCode < 58)
+							|| (charCode > 95 && charCode < 106)
+							|| (charCode == 8) || (charCode == 9)
+							|| (charCode == 12) || (charCode == 27)
+							|| (charCode == 37) || (charCode == 39)
+							|| (charCode == 46)) {
+						return true;
+					}
+
+					return false;
+				}
+			});
 	
 });
 
 
 function showAddRootDomainView(){
 	var table = "";
-	table += "<table style='margin-top:15px;margin-left:20px;width:100%;' class='imagetable' id='tableId'>";
-	table += "<tr id='0' pid='0'><td class='iconWidth'><img  class='expandDomain' src='images/expand.png'/></td><td class='iconWidth'><img style='display:none' class='newDomain' src='images/new.png'/></td><td class='iconWidth'><img class='saveDomain' src='images/save.png'/></td><td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td><td style='border-bottom:none'><input class='autoCompleteWiki' size='30' type='text'></td><td></td><td></td><td></td></tr>"
+	table += "<table style='margin-top:15px;width:100%;' class='imagetable' id='tableId'>";
+	table += "<tr id='0' pid='0'><td class='iconWidth'><img style='display:none' class='expandDomain' src='images/expand.png'/></td>" +
+			"<td class='iconWidth'><img style='display:none' class='newDomain' src='images/new.png'/></td>" +
+			"<td class='iconWidth'><img class='saveDomain' src='images/save.png'/></td>" +
+			"<td class='iconWidth'><img class='deleteDomain' src='images/cross.png'/></td>" +
+			"<td class='titleClass'><span class='spanTitle'></span><input class='autoCompleteWiki' size='30' type='text'></td>" +
+			"<td></td><td></td><td></td></tr>"
 	$('#domainDivId').prepend(table);
 }
 
 
 
-function updateDomain(key,parentKey,title,weightage,row) {
+function saveDomain(key,parentKey,title,weightage,row) {
 	if(weightage==undefined){
 		weightage = '100';
 	}
@@ -270,7 +315,7 @@ function updateDomain(key,parentKey,title,weightage,row) {
 			title : title,
 			weightage : weightage
 		};
-	var url ='updateDomain';
+	var url ='saveOrUpdateDomain';
 	$.ajax({
 		type : 'POST',
 		url : url,
@@ -293,7 +338,8 @@ function deleteDomain(key,parentKey) {
 	var data = {
 			key : key,
 			parentKey : parentKey,
-		};
+		}
+	
 	var url ='deleteDomain';
 	$.ajax({
 		type : 'POST',
